+++ packages/core/src/observer.ts
@@ -6,24 +6,24 @@ export class Observer {
 
   public static adapters: Adapters;
   public static interfaces: string[] = [];
-  // public static get interfaces() {
-  //   return Object.keys(Observer.adapters);
-  // }
   public static rootInterface: Root;
 
-  public static updateOptions(options: Options) {
+  public static updateOptions(options: Partial<Options>) {
     if (!options.adapters) {
       throw new Error('adapters are required!');
     }
 
-    if (!options.rootInterface) {
-      throw new Error('rootInterface is required!');
-    }
     if (options.adapters) {
-      Observer.adapters = options.adapters;
+      Observer.adapters = Utils.concat(false, Observer.adapters, options.adapters);
       Observer.interfaces = Object.keys(Observer.adapters);
     }
+
     Observer.rootInterface = options.rootInterface || Observer.interfaces[0];
+
+    if (!Observer.rootInterface) {
+      throw new Error('rootInterface is required!');
+    }
+
   }
 
   /**