+++ packages/core/src/services/adapter.service.ts
@@ -1,5 +1,6 @@
 import { Adapter, Adapters, ModuleElementType } from '../interfaces';
 import { ModuleElementService } from './module-element.service';
+import { Observer } from '../observer';
 
 export class AdapterService extends ModuleElementService {
 
@@ -23,6 +24,8 @@ export class AdapterService extends ModuleElementService {
       throw new Error('Adapter name not found!');
     }
     this.elements[name] = adapter;
+    const options = {adapters: this.elements as any as Adapters};
+    Observer.updateOptions(options);
     return this.elements;
   }
 }