+++ packages/shopify-tda/src/services/locales.service.ts
@@ -6,12 +6,12 @@ export class LocalesService extends LocalesRestService {
   public static getInstance() {
     return this.instance;
   }
-  constructor() {
+  constructor(doNotTranslateDefaultLanguage: boolean = false, showMissingTranslation: boolean = false) {
     let url = `${BaseService.baseUrl}/shopify/api/themes/${(window as any).Shopify.theme.id}/locales`;
     if ((window as any).Shopify.shop) {
       url += `?shop=${(window as any).Shopify.shop}`;
     }
-    super(url, true);
+    super(url, doNotTranslateDefaultLanguage, showMissingTranslation);
     if (LocalesService.instance) {
       return LocalesService.instance;
     }