+++ packages/shopify/src/formatters/money.formatter.ts
@@ -25,10 +25,9 @@ export const moneyFormatter: Formatter = {
 
     if (typeof cents === 'string') {
       cents = cents.replace('.', '');
+      // cents to float number
+      cents = parseFloat(cents.toString());
     }
-
-    // cents to float number
-    cents = parseFloat(cents.toString());
 
     const matchedFormat = formatString.match(placeholderRegex);